chore: rework methods files and exports
